backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Google AI
google-generativeai==0.3.2

# Audio Processing
SpeechRecognition==3.10.1
static-ffmpeg==2.5
pydub==0.25.1

# NLP Utilities  
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Testing & Evaluation
pytest==7.4.4
pytest-asyncio==0.23.3

# Logging
structlog==24.1.0

# Development
black==24.1.1
isort==5.13.2
